move CIP to a separate file
